Added login functionality, (4.21)

+++ part4/blogs/controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogsRouter = require('express').Router()
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
@@ -8,7 +9,10 @@ blogsRouter.get('/', async (request, response) => {
 })
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
+  if (!body.title || !body.url) return response.status(400).json({ error: 'title or url missing' })
+  const user = await User.findById(decodedToken.id)
   if (!user) return response.status(400).json({ error: 'User not found' })
   const blog = new Blog({
     title: request.body.title,
@@ -29,6 +33,11 @@ blogsRouter.get('/:id', async (request, response) => {
   response.status(404).end()
 })
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log(decodedToken)
+  if (!decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== decodedToken.id) return response.status(401).json({ error: 'only the creator can delete a blog' })
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
